Only consider extension tables when joining on pg_extension tables
We have some tables which directly or indirectly depend on `intarray`.
Because there's tables within `extconfig`, the subselect in the `SELECT`
clause yields a `NULL` while the `extension` is non-null. As a result,
seldump was skipping the tables as (presumably) it assumed these
table are extension configuration tables, and not tables which simply
depend on this extension

This commit adds a join which checks that the table we're considering is
actually an extension configuration table, and not a "regular" table that has an
extension as a dependency.